src/requirements.txt
asn1crypto==0.24.0
blinker==1.4
botocore==1.12.83
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.5
docutils==0.14
Flask==1.0.2
Flask-Email==1.4.4
Flask-Mail==0.9.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
pynamodb==3.3.3
python-dateutil==2.7.5
python-dotenv==0.10.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
